requirement.txt
torch==1.4.0
torchvision==0.5.0
nltk==3.5
spacy==2.3.0
einops
tqdm
pudb
ipdb
pillow
h5py
pandas
